Renamed function icon file name
